Задание 5.14
